Don't use bufio.Scanner to read the log
While using the scanner is convenient as it abstracts away EOL handling,
it forces the concept of tokens. Tokens have to fit into a particular
buffer size, and when logs have tons of uninterrupted text, this causes
the program to fail.

This commit replaces the use of bufio.Scanner with bufio.Reader.

Signed-off-by: Eric Chlebek <eric@sensu.io>